+++ crates/cwl-execution/src/runner.rs
@@ -19,7 +19,7 @@ use crate::{
 use commonwl::{
     Argument, CWLDocument, CWLType, Command, CommandLineTool, DefaultValue, Directory, Entry, File, PathItem, ScatterMethod, SingularPlural,
     StringOrDocument, StringOrNumber, Workflow,
-    inputs::CommandLineBinding,
+    inputs::{CommandLineBinding, LinkMerge},
     requirements::{DockerRequirement, InlineJavascriptRequirement, Requirement, StringOrInclude},
 };
 use log::{info, warn};
@@ -96,7 +96,7 @@ pub fn run_workflow(
 
                 //try input
                 if let Some(input) = workflow.inputs.iter().find(|i| i.id == *source) {
-                    let value = evaluate_input(input, &input_values.inputs)?;                    
+                    let value = evaluate_input(input, &input_values.inputs)?;
                     match value {
                         DefaultValue::Any(val) if val.is_null() => continue,
                         _ => {
@@ -105,16 +105,24 @@ pub fn run_workflow(
                     }
                 }
 
-                if workflow.has_requirement(Requirement::MultipleInputFeatureRequirement) && source.starts_with("["){
+                if workflow.has_requirement(Requirement::MultipleInputFeatureRequirement) && source.starts_with("[") {
                     //source can be array of input IDs if this requirement is set!
                     let array: Vec<String> = serde_yaml::from_str(source)?;
                     let mut data = vec![];
                     for item in array {
                         if let Some(input) = workflow.inputs.iter().find(|i| i.id == item) {
                             let value = evaluate_input(input, &input_values.inputs)?;
-                            data.push(value);
-                        }
-                        else {
+                            match parameter.link_merge {
+                                None | Some(LinkMerge::MergeNested) => data.push(value),
+                                Some(LinkMerge::MergeFlattened) => {
+                                    if let DefaultValue::Array(vec) = value {
+                                        data.extend(vec);
+                                    } else {
+                                        return Err("Expected array for MergeFlattened".into());
+                                    }
+                                }
+                            }
+                        } else {
                             return Err(format!("Could not find input: {item}").into());
                         }
                     }